+++ ufile/ufile-client-java/src/main/java/cn/ucloud/ufile/api/object/GetStreamApi.java
@@ -39,7 +39,7 @@ public class GetStreamApi extends UfileObjectApi<DownloadStreamBean> {
     /**
      * 流读取的buffer大小，Default = 256 KB
      */
-    private long bufferSize = UfileConstants.DEFAULT_BUFFER_SIZE;
+    private int bufferSize = UfileConstants.DEFAULT_BUFFER_SIZE;
 
     /**
      * 构造方法
@@ -86,7 +86,7 @@ public class GetStreamApi extends UfileObjectApi<DownloadStreamBean> {
      * @param bufferSize Buffer大小
      * @return {@link GetStreamApi}
      */
-    public GetStreamApi setBufferSize(long bufferSize) {
+    public GetStreamApi setBufferSize(int bufferSize) {
         this.bufferSize = bufferSize;
         return this;
     }
